+++ db/migrate/20231229085758_remove_redundant_fields.rb
@@ -1,0 +1,10 @@
+class RemoveRedundantFields < ActiveRecord::Migration[7.1]
+  def change
+    safety_assured do
+      remove_column :teams_players, :form
+      remove_column :teams_players, :shirt_number
+      remove_column :players, :points
+      remove_column :players, :statistic
+    end
+  end
+end