+++ CHANGELOG.md
@@ -15,6 +15,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Modified
 - RefreshSelectedService
+- removed redundant db fields
 
 ### Fixed
 - checkbox clicking